Splittext as a component
