+++ app/shared/components/Typography/component.jsx
@@ -1,16 +1,8 @@
 import React from 'react'
 import Button from '../Button/component.jsx'
 import ButtonGroup from '../ButtonGroup/component.jsx'
+import SplitText from '../SplitText/component.jsx'
 
-const SplitText = ({text, wrapper, modifiers = 'heading-inverted-test', highlight}) => {
-  let str = text.split(' ')
-  let Wrapper = `${wrapper || 'h2'}`
-  str = str.map((item, i) => {
-    let high = highlight.indexOf(i) !== -1 ? 'highlighted' : null
-    return <span key={i} className={high}>{item} </span>
-  })
-  return <Wrapper className={modifiers}>{str}</Wrapper>
-}
 
 
 const Typography = props => {
